Added the curved bottom navigation and its relevant pages
