+++ lib/Mechanics.dart
@@ -1,11 +1,13 @@
 import 'package:flutter/material.dart';
 
 class MechanicsPage extends StatelessWidget {
+  const MechanicsPage({super.key});
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Mechanics'),
+        title: const Text('Mechanics'),
       ),
       body: Padding(
         padding: const EdgeInsets.all(16.0),
