+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   cupertino_icons: ^1.0.6
   smooth_page_indicator: ^1.1.0
   shared_preferences: ^2.2.3
+  curved_navigation_bar: ^1.0.6
 
 dev_dependencies:
   flutter_test:
